refactor to split component, reducer, style, and types in multiple files
