+++ packages/ui/src/Popup/Accounts/AccountView.tsx
@@ -2,7 +2,7 @@ import { networkLinks } from '@polymathnetwork/extension-core/constants';
 import { IdentifiedAccount } from '@polymathnetwork/extension-core/types';
 import { recodeAddress } from '@polymathnetwork/extension-core/utils';
 import { SvgCheck, SvgDotsVertical, SvgPencilOutline } from '@polymathnetwork/extension-ui/assets/images/icons';
-import { Option } from '@polymathnetwork/extension-ui/components/OptionSelector';
+import { Option } from '@polymathnetwork/extension-ui/components/OptionSelector/types';
 import React, { FC, useContext, useState } from 'react';
 import { useHistory } from 'react-router-dom';
