+++ packages/ui/src/components/OptionSelector/index.tsx
@@ -1,0 +1,5 @@
+import { memo } from 'react';
+
+import { OptionSelector as OptionSelectorComponent } from './OptionSelector';
+
+export const OptionSelector = memo(OptionSelectorComponent, () => true);
